requirements.txt
requests
beautifulsoup4
pytz
PyYAML
tqdm
click
pendulum
